pyproject.toml
[project]
name = "swingmotionpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash",
    "gunicorn",
    "flask",
    "psycopg2-binary",
    "flask-login",
    "pandas",
    "flask-session>=0.8.0",
    "dash-mantine-components>=2.0.0",
    "dash-iconify>=0.1.2",
    "bcrypt>=4.3.0",
]
